moving the endpoints from the civ and pydt over
